# openapi.yaml
openapi: 3.0.3
info:
  title: Beacon Backend API
  version: 1.1.0
paths:
...
  /me/bootstrap:
    post:
      summary: Create/update current user profile from auth context
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, role]
              properties:
                full_name:
                  type: string
                phone_number:
                  type: string
                role:
                  type: string
                  enum: [citizen, student]
      responses:
        "200":
          description: Profile upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /sos:
    post:
      summary: Create SOS thread and notify friends
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSosRequest"
      responses:
        "200":
          description: SOS created
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /admin/auth/signup:
    post:
      summary: Admin/personnel signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAuthSignupRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthValidationErrorResponse"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/auth/login:
    post:
      summary: Admin/personnel login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAuthLoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthValidationErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Account is deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /incidents:
    post:
      summary: Create incident report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentRequest"
      responses:
        "201":
          description: Incident created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreateIncidentRequest:
      type: object
      required: [incident_type, description]
      properties:
        incident_type:
          type: string
        description:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        address:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
...
    UserDto:
      type: object
      required: [id, firebase_uid, email, full_name, role]
      properties:
        id:
          type: integer
        firebase_uid:
          type: string
        email:
          type: string
          format: email
          nullable: true
        full_name:
          type: string
        phone_number:
          type: string
          nullable: true
        role:
          type: string
          enum: [citizen, student]
        profile_image_url:
          type: string
          format: uri
          nullable: true
        status:
          type: string
          enum: [active, deactivated]
...
    CreateSosRequest:
      type: object
      required: [category]
      properties:
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        address:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        category:
          type: string
          enum: [medical, fire, violence, unknown]
...
    AdminAuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AdminAuthSignupRequest:
      type: object
      required: [full_name, email, password, role]
      properties:
        full_name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [personnel]
    AdminIdentity:
      type: object
      required: [id, full_name, email, role]
      properties:
        id:
          type: integer
        full_name:
          type: string
        email:
          type: string
        role:
          type: string
        role_id:
          type: integer
        permissions:
          type: array
          items:
            type: string
    AdminAuthResponse:
      type: object
      required: [token, admin]
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminIdentity"
...
    AdminAuthValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...